+++ test/drab/templates_test.exs
@@ -1,12 +1,12 @@
-defmodule Drab.TemplatesTest do
+defmodule Drab.TemplateTest do
   use ExUnit.Case, ascync: true
-  import Drab.Templates
-  doctest Drab.Templates
+  import Drab.Template
+  doctest Drab.Template
 
   test "render templates" do
-    assert String.contains?(render_template("call.alert.button.ok.html.eex", [label: "MYLABEL"]), "MYLABEL")
-    assert String.contains?(render_template("call.alert.button.cancel.html.eex", [label: "MYLABEL"]), "MYLABEL")
-    assert String.contains?(render_template("call.alert.html.eex", 
+    assert String.contains?(render_template("modal.alert.button.ok.html.eex", [label: "MYLABEL"]), "MYLABEL")
+    assert String.contains?(render_template("modal.alert.button.cancel.html.eex", [label: "MYLABEL"]), "MYLABEL")
+    assert String.contains?(render_template("modal.alert.html.eex", 
       [title: "TITLE", class: "CLASS", body: "BODY", buttons: "buttons html"]), "TITLE")
   end